requirements.txt
ts3==2.0.0b3
Django==3.1.*
djangorestframework==3.*
django-environ
